pulled functions into utils prettier
