+++ ai_genomics/pipeline/patent_data/get_ai_genomics_patents.py
@@ -3,66 +3,14 @@ AI patent ids and identify patent ids that have
 genomics-related classification codes."""
 #################################################
 from ai_genomics import bucket_name, config
-from ai_genomics.utils.patent_data.est_conn import est_conn
+from ai_genomics.utils.patent_data.get_ai_genomics_patents_utils import est_conn
 from ai_genomics.getters.data_getters import s3, load_s3_data, save_to_s3
 
 import pandas as pd
 import numpy as np
+import time
 
 #################################################
-
-
-def clean_ipc_codes(genomics_codes: dict) -> list:
-    """Prepares IPC codes in format to query Google BigQuery.
-
-    Inputs:
-        genomics_codes (dict): Dictionary of patent classification codes relevant to genomics.
-    Outputs:
-        ipc_codes_clean (list): List of clean IPC codes relevant to genomics.
-    """
-    ipc_codes_clean = []
-
-    if "ipc" in genomics_codes.keys():
-        for ipc_code in genomics_codes["ipc"].keys():
-            if len(ipc_code) == 14:
-                if ipc_code[6] == "0":
-                    ipc_codes_clean.append(
-                        ipc_code[:4] + ipc_code[7:8] + "/" + ipc_code[8:10]
-                    )
-                else:
-                    ipc_codes_clean.append(
-                        ipc_code[:4] + ipc_code[6:8] + "/" + ipc_code[8:10]
-                    )
-            else:
-                ipc_codes_clean.append(ipc_code + "/")
-
-        return ipc_codes_clean
-    else:
-        print("no IPC code keys in genomics_codes dict!")
-
-
-def make_table_query(
-    class_sys: str, classification_codes: list
-) -> str:  # used to create bespoke genomics patent BigQuery tables
-    """Queries for BigQuery console to generate new tables.
-
-    Inputs:
-        classification_codes (list): list of classification codes.
-        class_sys: classification system to query.
-
-    Outputs:
-        q (str): string to query BigQuery console with.
-    """
-
-    class_str = "'" + "', '".join(classification_codes) + "'"
-
-    q = (
-        f"SELECT application_number "
-        f"FROM `patents-public-data.patents.publications`, UNNEST({class_sys}) AS {class_sys}__u "
-        f"WHERE {class_sys}__u.code IN ({class_str})"
-    )
-
-    return q
 
 
 def make_query_chunks(uspto_patent_ids, n_chunks: int, table: str) -> list:
@@ -93,9 +41,7 @@ def make_query_chunks(uspto_patent_ids, n_chunks: int, table: str) -> list:
     return uspto_queries
 
 
-def query_patent_data(
-    conn, query_chunks: list
-) -> pd.DataFrame:  
+def query_patent_data(conn, query_chunks: list) -> pd.DataFrame:
     """Queries genomics tables in query chunks.
     Args:
         conn: Google BigQuery connection.
@@ -120,6 +66,8 @@ def query_patent_data(
 
 if __name__ == "__main__":
     # load data
+    start = time.time()
+
     uspto_data = load_s3_data(bucket_name, config["uspto_file"])
     uspto_patent_ids = uspto_data[uspto_data.flag_patent == 1]["doc_id"]
     print("loaded data")
@@ -135,3 +83,5 @@ if __name__ == "__main__":
     print("queried genomics related patents")
     # save output
     save_to_s3(s3, bucket_name, ai_patents, config["ai_genomics_patents_file"])
+
+    print("It took", time.time() - start, "seconds.")
